Refactor / move files.
